+++ app.js
@@ -13,7 +13,8 @@ const registerRouter = require('./routes/register');
 const accountRouter = require('./routes/account');
 const demoRouter = require('./routes/demo');
 const logoutRouter = require('./routes/logout');
-const apiRouter = require('./routes/api');
+const apiListRouter = require('./routes/api-lists');
+const apiTaskRouter = require('./routes/api-tasks');
 
 const { sessionSecret } = require('./config');
 const { restoreUser } = require('./auth');
@@ -53,7 +54,8 @@ app.use('/register', registerRouter);
 app.use('/account', accountRouter);
 app.use('/demo', demoRouter);
 app.use('/logout', logoutRouter);
-app.use('/api', apiRouter);
+app.use('/api/lists', apiListRouter);
+app.use('/api/tasks', apiTaskRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
